+++ MOP/src/Rules/Configuration/ServerSources.cs
@@ -1,0 +1,7 @@
+﻿namespace MOP.Rules.Configuration
+{
+    class ServerSources
+    {
+        public string[] Sources { get; set; }
+    }
+}